Removed: Command line class Changed: Some tweaks Changed: Max item calculation is more accurate now Changed: to Version v0.30 Add: First implementation for jump drives Fixed: Calculation for line height for text Fixed: Pack algo in ValueType

+++ VisualInformationSystem/DataCollectorAirVent.cs
@@ -75,7 +75,7 @@ namespace IngameScript
                     dc_ = obj;
                 }
 
-                public override double indicator() => (double)dc_.pressurizeAble_ / dc_.Blocks.Count;
+                public override double indicator() => (double)dc_.pressurizeAble_ / (double)dc_.Blocks.Count;
                 public override ValueType min() => new ValueType(0.0);
                 public override ValueType max() => new ValueType(dc_.Blocks.Count);
                 public override ValueType value() => new ValueType();
